split parameter extraction logic into a separate file
